requirements.txt
oslo.messaging>=1.4.0
eventlet>=0.16.1
